feat: Integration tests
