+++ resources/scripts/pytest_otel/.gitignore
@@ -12,6 +12,8 @@ __pycache__
 **/*.egg-info
 **/build
 **/dist
+junit-*.xml
 otel-traces-file-output.json
 temp
 test_spans.json
+tests/tests.json
