+++ resources/scripts/pytest_otel/setup.cfg
@@ -5,7 +5,7 @@ long_description = file: README.md
 long_description_content_type = text/markdown
 url = https://github.com/elastic/apm-pipeline-library/tree/master/resources/scripts/pytest_otel
 maintainer = Ivan Fernandez Calvo
-version = 0.0.2
+version = 0.0.3
 license = Apache-2.0
 license_file = LICENSE.txt
 platforms = any
