+++ resources/scripts/pytest_otel/tox.ini
@@ -8,16 +8,18 @@ envlist =
 
 [testenv]
 deps =
-    pytest
+    pytest==6.2.5
+    pytest-docker==0.10.3
 commands =
     pytest {tty:--color=yes} --capture=no \
-      -p pytester --runpytest=subprocess \
-      --junitxml {toxworkdir}{/}junit.{envname}.xml
+        -p pytester --runpytest=subprocess \
+        --junitxml {toxworkdir}{/}junit-{envname}.xml \
+        tests/test_pytest_otel.py
 
 [testenv:linting]
 basepython = python3
 skip_install = true
 deps =
-    pre-commit>=1.11
+    pre-commit==2.16.0
 commands =
     pre-commit run
